+++ Service/ContactFormService.php
@@ -9,9 +9,9 @@
 
 namespace c975L\ContactFormBundle\Service;
 
-use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\Form\Form;
 use Symfony\Component\HttpFoundation\RequestStack;
+use c975L\ConfigBundle\Service\ConfigServiceInterface;
 use c975L\ServicesBundle\Service\ServiceToolsInterface;
 use c975L\ServicesBundle\Service\ServiceUserInterface;
 use c975L\ContactFormBundle\Entity\ContactForm;
@@ -29,10 +29,10 @@ use c975L\ContactFormBundle\Service\User\ContactFormUserInterface;
 class ContactFormService implements ContactFormServiceInterface
 {
     /**
-     * Stores container
-     * @var ContainerInterface
+     * Stores ConfigServiceInterface
+     * @var ConfigServiceInterface
      */
-    private $container;
+    private $configService;
 
     /**
      * Stores ContactFormEmailInterface
@@ -65,7 +65,7 @@ class ContactFormService implements ContactFormServiceInterface
     private $serviceUser;
 
     public function __construct(
-        ContainerInterface $container,
+        ConfigServiceInterface $configService,
         ContactFormEmailInterface $contactFormEmail,
         ContactFormFactoryInterface $contactFormFactory,
         RequestStack $requestStack,
@@ -73,7 +73,7 @@ class ContactFormService implements ContactFormServiceInterface
         ServiceUserInterface $serviceUser
     )
     {
-        $this->container = $container;
+        $this->configService = $configService;
         $this->contactFormEmail = $contactFormEmail;
         $this->contactFormFactory = $contactFormFactory;
         $this->request = $requestStack->getCurrentRequest();
@@ -146,7 +146,8 @@ class ContactFormService implements ContactFormServiceInterface
     public function isNotBot($username)
     {
         $bot = null === $this->request->getSession()->get('time');
-        $bot = $bot ? true : $this->request->getSession()->get('time') + $this->container->getParameter('c975_l_contact_form.delay') > time();
+        $bot = $bot ? true : $this->request->getSession()->get('time') + $this->configService->getParameter('c975LContactForm.delay') > time();
+
         $bot = $bot ? true : null !== $username;
 
         return ! $bot;