+++ ChangeLog.md
@@ -1,5 +1,23 @@
 # Changelog
 
+v2.0
+----
+**Upgrading from v1.x? Check UPGRADE.md**
+- Created branch 1.x (30/08/2018)
+- Made use of c975L/ConfigBundle (30/08/2018)
+- Added Route `contactform_config` (30/08/2018)
+- Removed declaration of parameters in Configuration class as they are end-user parameters and defined in c975L/ConfigBundle (30/08/2018)
+- Added `bundle.yaml` (30/08/2018)
+- Updated `README.md` (30/08/2018)
+- Added `UPGRADE.md` (30/08/2018)
+- Added `ContactFormVoter.php` (30/08/2018)
+- Added toolbar (30/08/2018)
+- Added Route `contactform_dashboard` (30/08/2018)
+
+
+v1.x
+====
+
 v1.10.5
 -------
 - Fixed missing break (30/08/2018)